requirements.txt
agno==1.8.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
beautifulsoup4==4.13.5
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
distro==1.9.0
docstring_parser==0.17.0
gitdb==4.0.12
GitPython==3.1.45
google==3.0.0
google-auth==2.40.3
google-genai==1.33.0
groq==0.31.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
narwhals==2.3.0
numpy==2.3.2
packaging==25.0
pandas==2.3.2
pillow==11.3.0
protobuf==6.32.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.5
rich==14.1.0
rpds-py==0.27.1
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
streamlit==1.49.1
streamlit-code-editor==0.1.22
tenacity==9.1.2
toml==0.10.2
tomli==2.2.1
tornado==6.5.2
typer==0.17.4
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
websockets==15.0.1
